+++ src/main/java/com/xytong/model/po/ShPO.java
@@ -15,12 +15,12 @@ public class ShPO implements Serializable {
     /**
      * 
      */
-    private Integer id;
+    private Long id;
 
     /**
      * 
      */
-    private Integer userFkey;
+    private Long userFkey;
 
     /**
      * 
@@ -47,28 +47,28 @@ public class ShPO implements Serializable {
     /**
      * 
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
     /**
      * 
      */
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
     /**
      * 
      */
-    public Integer getUserFkey() {
+    public Long getUserFkey() {
         return userFkey;
     }
 
     /**
      * 
      */
-    public void setUserFkey(Integer userFkey) {
+    public void setUserFkey(Long userFkey) {
         this.userFkey = userFkey;
     }
